requirements.txt
itsdangerous
Flask==0.10.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
pymongo==3.2.2
jsonpickle==0.9.3
pytest==2.9.1
python-slugify==1.2.0
bleach==1.4.2
freezegun==0.3.7
passlib==1.6.5
honcho==0.7.1
mongomock==3.3.0